fix error on server

+++ app.py
@@ -28,7 +28,7 @@ async def check_subscriptions():
                 FROM users u
                 JOIN payments p ON p.user_id = u.id
                 JOIN vpn_configs vc ON vc.payment_id = p.payment_id
-                WHERE vc.session_end::date = CURRENT_DATE + $1
+                WHERE vc.session_end::date = CURRENT_DATE + $1::int
                 AND NOT EXISTS (
                     SELECT 1 FROM notifications_queue nq
                     WHERE nq.payment_id = vc.payment_id
